fixes skill exploits an adds smelting drops mechanic, balance tweaks

+++ src/main/java/com/robertx22/age_of_exile/database/data/player_skills/ItemCraftExp.java
@@ -6,10 +6,10 @@ import net.minecraft.util.registry.Registry;
 
 public class ItemCraftExp {
 
-    public float exp = 0;
+    public int exp = 0;
     public String item_id = "";
 
-    public ItemCraftExp(float exp, Item item) {
+    public ItemCraftExp(int exp, Item item) {
         this.exp = exp;
         this.item_id = Registry.ITEM.getId(item)
             .toString();
